requirements.txt
alabaster==0.7.10
apturl==0.5.2
asn1crypto==0.24.0
Babel==2.5.1
backports.weakref==1.0rc1
bcrypt==3.1.4
beautifulsoup4==4.4.1
bleach==2.1.2
blinker==1.3
Brlapi==0.6.4
cache==1.0.3
cffi==1.11.2
chardet==2.3.0
checkbox-support==0.22
click==6.7
command-not-found==0.3
cryptography==2.1.4
cycler==0.10.0
Cython==0.27.3
decorator==4.2.1
defer==1.0.6
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
feedparser==5.1.3
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
get==0.0.38
guacamole==0.9.2
html5lib==1.0.1
httplib2==0.9.1
idna==2.6
imagesize==0.7.1
imgaug==0.2.5
ipykernel==4.8.0
ipyparallel==6.0.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
joblib==0.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
language-selector==0.1
Lasagne==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nose==1.3.7
notebook==5.3.1
numpy==1.14.0
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pandas==0.22.0
pandocfilters==1.4.2
paramiko==2.4.0
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==3.1.2
pkg-resources==0.0.0
plainbox==0.25
post==0.0.26
progressbar2==3.34.3
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
public==0.0.65
pyasn1==0.1.9
pycparser==2.18
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
-e git+https://github.com/lisa-lab/pylearn2.git@58ba37286182817301ed72b0f143a89547b3f011#egg=pylearn2
PyNaCl==1.2.1
pyparsing==2.0.3
python-apt==1.1.0b1
python-dateutil==2.6.1
python-debian==0.1.27
python-systemd==231
python-utils==2.2.0
pytz==2017.3
PyWavelets==0.5.2
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
query-string==0.0.28
reportlab==3.3.0
request==0.0.26
requests==2.9.1
scikit-image==0.13.1
scikit-learn==0.17
scipy==1.0.0
screen-resolution-extra==0.0.0
Send2Trash==1.4.2
sessioninstaller==0.0.0
simplegeneric==0.8.1
simplejson==3.8.1
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
SQLAlchemy==1.1.15
ssh-import-id==5.5
system-service==0.3
TA-Lib==0.4.10
tensorflow-gpu==1.4.1
tensorflow-tensorboard==0.4.0rc3
terminado==0.8.1
testpath==0.3.1
Theano==0.8.0
tornado==4.5.3
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.0
virtualenv==15.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.0
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
